passing tests, more refactors
