+++ assembler/src/lib.rs
@@ -6,3 +6,6 @@ mod parse;
 pub use assembler::*;
 
 pub use lib_rv32_common::constants;
+
+#[cfg(test)]
+mod test;
