+++ cli/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lib_rv32_mcu"
+name = "lib_rv32_cli"
 description = "Extension to lib_rv32 that provides an MCU implementation"
 keywords = ["riscv", "testing", "simulator", "emulator"]
 authors = ["Trevor McKay <tm@trmckay.com>"]
@@ -13,15 +13,10 @@ readme = "./README.md"
 [dependencies]
 serde_json = "1.0.*"
 lazy_static = "1.4.*"
-log = "0.4.*"
 clap = "2.33.*"
-lib_rv32_sim = { path = "../isa-sim" }
+log = "0.4.*"
+lib_rv32_mcu = { path = "../mcu" }
 lib_rv32_asm = { path = "../assembler" }
-lib_rv32_common = { path = "../common" }
-
-[dev-dependencies]
-glob = "0.3.*"
-
 
 [[bin]]
 name = "lib-rv32-cli"
