--- cli/tests/programs/mul/test_case.json
@@ -1,9 +1,0 @@
-{
-    "max_cycles": 80,
-    "stop_pc": 24,
-    "assertions": {
-        "registers": {
-            "a0": 20
-        }
-    }
-}
