+++ common/src/bits.rs
@@ -12,7 +12,7 @@ macro_rules! bit {
 /// Example:
 ///
 /// ```
-/// # use lib_rv32::bit_slice;
+/// # use lib_rv32_common::bit_slice;
 /// bit_slice!(0b1101, 3, 2) == 0b11;
 /// bit_slice!(0b1101, 1) == 0b0;
 /// ```
@@ -32,7 +32,7 @@ macro_rules! bit_slice {
 /// Example:
 ///
 /// ```
-/// # use lib_rv32::bit_concat;
+/// # use lib_rv32_common::bit_concat;
 /// bit_concat!(
 ///     (0b111, 3),
 ///     (0b01, 2)
@@ -56,7 +56,7 @@ macro_rules! bit_concat {
 /// Example:
 ///
 /// ```
-/// # use lib_rv32::bit_extend;
+/// # use lib_rv32_common::bit_extend;
 /// bit_extend!(0b1, 8) == 0b1111_1111;
 /// ```
 #[macro_export]
