+++ common/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod bits;
 pub mod constants;
 pub mod util;
+pub mod instructions;
