+++ mcu/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lib_rv32_isa"
+name = "lib_rv32_mcu"
 description = "A library and CLI tool for emulating, testing, and learning RISC-V."
 keywords = ["riscv", "testing", "simulator", "emulator"]
 authors = ["Trevor McKay <tm@trmckay.com>"]
@@ -12,8 +12,12 @@ readme = "./README.md"
 
 [dependencies]
 log = "0.4.*"
-lib_rv32_common = { path = "../common" }
+lib_rv32_isa = { path = "../isa-sim" }
+
+[dev-dependencies]
+serde_json = "1.0.*"
+glob = "0.3.*"
 
 [lib]
-name = "lib_rv32_isa"
+name = "lib_rv32_mcu"
 path = "src/lib.rs"
