+++ mcu/programs/mul/init.s
