+++ mcu/src/register_file.rs
@@ -1,6 +1,5 @@
-pub use lib_rv32_sim::traits::RegisterFile as RegisterFileTrait;
-use lib_rv32_sim::RiscvError;
-use lib_rv32_common::constants::*;
+pub use lib_rv32_isa::traits::RegisterFile as RegisterFileTrait;
+use lib_rv32_isa::{RiscvError, common::constants::*};
 use log::info;
 
 /// Heap allocated implementation of a register file.
